requirements-test.txt
coverage>=4.4.1
mock>=1.0.1
flake8>=2.1.0
codecov>=2.0.0
flake8-future-import>=0.4.5
django-functest==1.5.3
check-manifest
isort
testfixtures
pytest
pytest-django
